extract http status checks
